add test functionality for document adding
